refactor[elastoplasticity]: renamed residual functions to internal body force, removed redundant checks, and cleaned Newton solver logic
